move friendship operations into separate logic
