+++ src/main/java/ru/yandex/practicum/filmorate/controller/FriendShipController.java
@@ -1,0 +1,56 @@
+package ru.yandex.practicum.filmorate.controller;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.model.FriendShip;
+import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.service.FriendShipService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/users")
+@Slf4j
+@Validated
+public class FriendShipController {
+
+    private final FriendShipService friendShipService;
+
+    @Autowired
+    public FriendShipController(FriendShipService friendShipService) {
+        this.friendShipService = friendShipService;
+    }
+
+    @PutMapping("/{userId}/friends/{friendId}")
+    @ResponseStatus(HttpStatus.OK)
+    public FriendShip addFriend(
+            @PathVariable("userId") Integer userId,
+            @PathVariable("friendId") Integer friendId
+    ) {
+        return friendShipService.addFriend(userId, friendId);
+    }
+
+    @GetMapping("/{userId}/friends")
+    @ResponseStatus(HttpStatus.OK)
+    public List<User> friendGet(@PathVariable("userId") Integer userId) {
+        return friendShipService.friendGet(userId);
+    }
+
+    @DeleteMapping("/{userId}/friends/{friendId}")
+    public void removeUserFromFriends(
+            @PathVariable Integer userId,
+            @PathVariable Integer friendId
+    ) {
+        friendShipService.removeUserFromFriends(userId, friendId);
+    }
+
+    @GetMapping("/{userId}/friends/common/{otherUserId}")
+    public List<User> getCommonFriends(
+            @PathVariable("userId") Integer userId,
+            @PathVariable("otherUserId") Integer otherUserId) {
+        return friendShipService.getCommonFriends(userId, otherUserId);
+    }
+}
