+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -7,12 +7,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.yandex.practicum.filmorate.model.FriendShip;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
 import java.util.Collection;
-import java.util.List;
 
 @RestController
 @RequestMapping("/users")
@@ -48,35 +46,5 @@ public class UserController {
     @GetMapping("/{id}")
     public User getUserById(@PathVariable("id") @Positive Integer id) {
         return userService.getUserById(id);
-    }
-
-    @GetMapping("/{userId}/friends")
-    @ResponseStatus(HttpStatus.OK)
-    public List<User> friendGet(@PathVariable("userId") Integer userId) {
-        return userService.friendGet(userId);
-    }
-
-    @DeleteMapping("/{userId}/friends/{friendId}")
-    public void removeUserFromFriends(
-            @PathVariable Integer userId,
-            @PathVariable Integer friendId
-    ) {
-        userService.removeUserFromFriends(userId, friendId);
-    }
-
-    @GetMapping("/{userId}/friends/common/{otherUserId}")
-    public List<User> getCommonFriends(
-            @PathVariable("userId") Integer userId,
-            @PathVariable("otherUserId") Integer otherUserId) {
-        return userService.getCommonFriends(userId, otherUserId);
-    }
-
-    @PutMapping("/{userId}/friends/{friendId}")
-    @ResponseStatus(HttpStatus.OK)
-    public FriendShip addFriend(
-            @PathVariable("userId") Integer userId,
-            @PathVariable("friendId") Integer friendId
-    ) {
-        return userService.addFriend(userId, friendId);
     }
 }
