+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -4,8 +4,6 @@ import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
 
 @Component
 public interface UserStorage {
@@ -15,13 +13,5 @@ public interface UserStorage {
 
     User updateUser(User user);
 
-    Optional<User> getUserById(Integer id);
-
-    void addFriend(Integer userId, Integer friendId);
-
-    void removeUserFromFriends(Integer id, Integer friendId);
-
-    List<User> getCommonFriends(Integer userId, Integer otherUserId);
-
-    List<User> friendGet(Integer userId);
+    User getUserById(Integer id);
 }
